Reduce the total number of Break and Continue Statements
Content: This issue was meant to reduce the total number of Break and
Continue statements in the code mapped during the sonar cube detection
for bugs and error. A total of 18 instances were presented our of which
few instances required the scope of including continue statements due to
empty conditions and functions which need to maintain the flow of the
code. Cases in the Switch conditions need to have break statements as
there is no explicit control specified after user opts for any such
case.

Note: few instances which had empty conditions were studied and left out
without removing the continue statements. Break statements which were
detected in the code smells are few and are required to be present for
the respective cases.