Remove the parentheses around the " *parameter* "
Content: In general, parentheses are used to determine any structural
specifications or a particular attention towards any action on the data.
In this issue the requirement to remove the parentheses is arisen due to
the unnecessary inclusion in the first place while coding. Parameters
declared can be utilized to specify or complete the function, redundant
parentheses are not encouraged while practicing good coding.