Add a default case to this switch.
Content: A default switch case is added to the code to provide a base
decision for the program to take an input. adding default cases allows
the system to take values even when the user does not chooses one making
the execution of the program smoothly. It is a good coding practice to
add default cases to a switch conditional statement.