Extract this nested try block into a separate method.
Content: Having Distinguished try blocks improves the functioning of the
in a better way. It increases readability and understanding of the code
to other developers or systems. Disadvantage of having nested try-catch
blocks is that the inner catch blocks may or may not catch the
exceptions from within. Avoiding this situation rather than experiencing
is suggested for a better coding environment and good development.